requirements.txt
# DCA Bot — Python dependencies
# Install with: pip install -r requirements.txt

python-binance==1.0.19
apscheduler==3.10.4
python-dotenv==1.0.0
pyyaml==6.0.1
pandas==2.1.0
tabulate==0.9.0
